LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-align-baseline-multicol-003-expected.txt










































FAIL .target > * 1 assert_equals:
<div data-offset-x="25"><span></span></div>
offsetLeft expected 25 but got 24
PASS .target > * 2
FAIL .target > * 3 assert_equals:
<div data-offset-x="95"><span></span></div>
offsetLeft expected 95 but got 311
PASS .target > * 4
FAIL .target > * 5 assert_equals:
<div data-offset-x="30"><span></span></div>
offsetLeft expected 30 but got 24
PASS .target > * 6
FAIL .target > * 7 assert_equals:
<div data-offset-x="135"><span></span></div>
offsetLeft expected 135 but got 50
PASS .target > * 8
FAIL .target > * 9 assert_equals:
<div data-offset-x="25"><span></span></div>
offsetLeft expected 25 but got 24
PASS .target > * 10
FAIL .target > * 11 assert_equals:
<div data-offset-x="140"><span></span></div>
offsetLeft expected 140 but got 170
PASS .target > * 12
